Project is active again. Started improving and optimizing code, although much more is hopefully to come, as I try to bring in all I have learnt from March until now to improve performance. Currently working on solving the tendency of the algorithm to simply fold all the time, which is not necessarily bad behavior, but certainly not very interesting, or how pro poker players play...

+++ headers/holdem/Player.h
@@ -40,14 +40,14 @@ public:
      * @param ante Amount bet
      * @return Amount anted (will be different to the required ante if player went all in)
      */
-    int anteUp(int ante);
+    unsigned long anteUp(unsigned long ante);
 
     /**
      * Increases the players money, when a round is won.
      *
      * @param winnings Amount won
      */
-    void collectWinnings(int winnings);
+    void collectWinnings(unsigned long winnings);
 
     int getMoney();
